vmodel: feature extraction + verify harness (2000 rows @ <1e-5)
src/ent/static_leaves.{h,c}: load CSW24 leave-value tables (short_csv =
L1..L6 leaves, full_csv = L7 racks). Binary-search lookup keyed on
canonical-form (blanks first, then alpha) tile indices.

src/ent/vmodel_features.{h,c}: C port of v_model_features.py
extract_features. Mirrors the family ordering exactly: [diff?] +
[static_leave?] + leave_tile (1-hot L=1 or count L>=3) + leave_pair
counts + [bag_exp] + [bag_eq indicators] + [exch_tile K1]. Terminal-
zero variant: leave_pts indicators + linear leave_pts + K1
drawn_indicators + drawn_pts_expected. Returns -1 on schema mismatch
(should never happen — fail loud).

vmodel_predict() bundles stratum/bucket lookup + feature extract + dot
product + sigmoid into a single call.

test/vmodel_test.c:test_vmodel_verify: reads model + Python gold CSV
(rack,leave,kind,diff,turn,bag,win_pct) + leave CSVs, runs C inference
on every row, asserts |Δ| < 1e-5 against Python. All 2000 stratified
sample rows pass at max |Δ| = 5.36e-7 — pure float32 rounding noise.

Float-precision note: drawn_indicators must be double (not float)
because Python rounds to 4 decimal places then exact-equality compares;
float32 round-trip loses the precision and the indicator misfires.
Coefs stay float32 (multiplied, not compared).

+++ src/ent/vmodel.h
@@ -37,7 +37,9 @@ typedef struct VBucket {
     int     n_pts;
     int    *pts;            // [n_pts]  optional pts_indicators
     int     n_drawn;
-    float  *drawn;          // [n_drawn]  optional drawn_indicators
+    double *drawn;          // [n_drawn]  optional drawn_indicators (double:
+                            // Python rounds to 4dp and compares exact-equal;
+                            // float32 would round-trip-lose precision)
     int     n_bag;
     uint8_t *bag_tiles;     // [n_bag]  python-canonical tile index (0='?'..26='Z')
     int8_t  *bag_counts;    // [n_bag]
